+++ src/Strategy/Node/Npm/PackageLock.hs
@@ -28,12 +28,13 @@ import Data.Foldable (asum, traverse_)
 import Data.Functor (void)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
-import Data.Maybe (isNothing, mapMaybe)
+import Data.Maybe (isNothing)
 import Data.Set (Set)
 import Data.Set qualified as Set
 import Data.Tagged (unTag)
-import Data.Text (Text, splitOn)
+import Data.Text (Text)
 import Data.Text qualified as Text
+import Data.Text.Extra qualified as TE
 import DepTypes (
   DepEnvironment (EnvDevelopment, EnvProduction),
   DepType (NodeJSType),
@@ -132,20 +133,12 @@ data NpmDepVertexLabel = NpmDepVertexEnv DepEnvironment | NpmDepVertexLocation T
 -- downloaded to @node_modules@. This function will adjust map keys to be names
 -- like in the @dependencies@ key by stripping out path components besides the final one..
 --
--- It also eliminates any keys which represent a nested path
--- e.g. @node_modules\/foo\/node_modules/bar@. This nesting happens when there
--- are multiple versions of the same package in the dependency tree of the
--- @package-lock.json@ file because one of the versions gets vendored.
+-- When npm installs a dep inside of another dep (for version conflicts) we get the string
+-- @node_modules/a/node_modules/b@.  We actually don't want to do sub-package matching here,
+-- so we only drop the first @node_modules/@, and if it has a sub-package, then we just
+-- won't ever query for that key.
 packagePathsToNames :: Map Text a -> Map Text a
-packagePathsToNames =
-  Map.fromList
-    . mapMaybe fixName
-    . Map.toList
-  where
-    fixName :: (Text, a) -> Maybe (Text, a)
-    fixName (k, v) = case filter (/= "node_modules") . splitOn "/" $ k of
-      [k'] -> Just (k', v)
-      _ -> Nothing
+packagePathsToNames = Map.mapKeys (TE.dropPrefix "node_modules/")
 
 buildGraph :: PkgLockJson -> Set Text -> Graphing Dependency
 buildGraph packageJson directSet =
@@ -170,8 +163,16 @@ buildGraph packageJson directSet =
     -- direct dependency in 'maybeAddDep'.
     addPeerDeps :: Has NpmGrapher sig m => NpmDepVertex -> m ()
     addPeerDeps currentPkg =
-      maybe (pure ()) graphPeerDeps (Map.lookup (lockName currentPkg) pkgJsonPackages)
+      traverse_ graphPeerDeps (Map.lookup (lockName currentPkg) pkgJsonPackages)
       where
+        graphPeerDeps :: Has NpmGrapher sig m => PkgLockPackage -> m ()
+        graphPeerDeps =
+          -- ignore the version range specified in "peerDependencies", we'll use
+          -- the resolved one from the main list of dependencies.
+          traverse_ (addNodeAndEdge . fst)
+            . Map.toList
+            . pkgPeerDeps
+
         addNodeAndEdge :: Has NpmGrapher sig m => Text -> m ()
         addNodeAndEdge peerDepName =
           case Map.lookup peerDepName packageDeps of
@@ -184,14 +185,6 @@ buildGraph packageJson directSet =
               -- It will be set as direct/deep and labeled then.
               edge currentPkg $ NpmDepVertex peerDepName (depVersion npmDep)
             Nothing -> pure ()
-
-        graphPeerDeps :: Has NpmGrapher sig m => PkgLockPackage -> m ()
-        graphPeerDeps =
-          -- ignore the version range specified in "peerDependencies", we'll use
-          -- the resolved one from the main list of dependencies.
-          traverse_ (addNodeAndEdge . fst)
-            . Map.toList
-            . pkgPeerDeps
 
     -- If not resolved, then likely a workspace dep, should be ignored.
     -- isRecursive lets us know if we are parsing a top-level or nested dep.